Split directory entry code #5
